+++ src/app/App.tsx
@@ -13,7 +13,6 @@ import { WelcomeScreen } from 'src/features/onboarding/welcome/WelcomeScreen'
 import { EnterPincodeScreen } from 'src/features/pincode/EnterPincodeScreen'
 import { SendConfirmationScreen } from 'src/features/send/SendConfirmationScreen'
 import { SendFormScreen } from 'src/features/send/SendFormScreen'
-import { WalletHome } from 'src/features/wallet/WalletHome'
 
 export class App extends Component {
   render() {
@@ -35,7 +34,7 @@ export class App extends Component {
           <Route path="pin" element={<EnterPincodeScreen />} />
 
           {/* TODO: For Splashscreen Testing */}
-          <Route path="wallet" element={<WalletHome />} />
+          <Route path="home" element={<HomeScreen />} />
           <Route path="loading" element={<Loading />} />
         </Routes>
       </BrowserRouter>
