+++ src/features/home/HeaderSectionEmpty.tsx
@@ -4,7 +4,7 @@ import { Color } from 'src/styles/Color'
 import { Font } from 'src/styles/fonts'
 import { Stylesheet } from 'src/styles/types'
 
-export function WalletEmpty() {
+export function HeaderSectionEmpty() {
   return (
     <Box direction="column">
       <h1 css={style.header}>Welcome to your Celo wallet!</h1>
@@ -16,8 +16,8 @@ export function WalletEmpty() {
         </Box>
         <p css={Font.body}>
           To fund your account you can <a href="#">buy currency</a> from an exchange,{' '}
-          <a href="#">redeem an invite code</a>, or ask a friend on Celo to{' '}
-          <a href="#">send a payment</a> to your address.
+          <a href="#">redeem an invite code</a>, or ask a friend on Celo to send a payment to{' '}
+          <a href="#">your address.</a>
         </p>
       </Box>
     </Box>
